streaming-to-cassandra/requirements.txt
apache-airflow
kafka-python
cassandra-driver
spark pyspark